[PB-1638] refactor DriveView breadcrumbs
