+++ src/app/backups/views/BackupsView/BackupsView.tsx
@@ -2,7 +2,8 @@ import { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet-async';
 import DeviceList from '../../components/DeviceList/DeviceList';
 import { Device } from '../../types';
-import Breadcrumbs, { BreadcrumbItemData } from 'app/shared/components/Breadcrumbs/Breadcrumbs';
+import Breadcrumbs from 'app/shared/components/Breadcrumbs/Breadcrumbs';
+import { BreadcrumbItemData } from 'app/shared/components/Breadcrumbs/types';
 import { backupsActions, backupsThunks } from 'app/store/slices/backups';
 import { useAppDispatch, useAppSelector } from 'app/store/hooks';
 import { DriveFolderData } from '@internxt/sdk/dist/drive/storage/types';
