+++ src/app/share/views/SharedLinksView/SharedView.tsx
@@ -31,7 +31,8 @@ import errorService from '../../../core/services/error.service';
 import ShareDialog from '../../../drive/components/ShareDialog/ShareDialog';
 import Avatar from '../../../shared/components/Avatar';
 import { AdvancedSharedItem, OrderBy, PreviewFileItem, SharedNamePath, UserRoles } from '../../../share/types';
-import Breadcrumbs, { BreadcrumbItemData } from '../../../shared/components/Breadcrumbs/Breadcrumbs';
+import Breadcrumbs from '../../../shared/components/Breadcrumbs/Breadcrumbs';
+import { BreadcrumbItemData } from 'app/shared/components/Breadcrumbs/types';
 import { getItemPlainName } from '../../../../app/crypto/services/utils';
 import Button from '../../../shared/components/Button/Button';
 import storageThunks from '../../../store/slices/storage/storage.thunks';
