+++ src/app/shared/components/Breadcrumbs/Breadcrumbs.tsx
@@ -2,17 +2,7 @@ import { CaretRight, DotsThree } from '@phosphor-icons/react';
 import { forwardRef, ReactNode } from 'react';
 import Dropdown from '../Dropdown';
 import BreadcrumbsItem from './BreadcrumbsItem/BreadcrumbsItem';
-
-export interface BreadcrumbItemData {
-  id: number;
-  label: string;
-  icon: JSX.Element | null;
-  active: boolean;
-  isFirstPath?: boolean;
-  dialog?: boolean;
-  isBackup?: boolean;
-  onClick?: () => void;
-}
+import { BreadcrumbItemData } from './types';
 
 interface BreadcrumbsProps {
   items: BreadcrumbItemData[];
