+++ src/app/shared/components/Breadcrumbs/types.tsx
@@ -1,0 +1,10 @@
+export interface BreadcrumbItemData {
+  id: number;
+  label: string;
+  icon: JSX.Element | null;
+  active: boolean;
+  isFirstPath?: boolean;
+  dialog?: boolean;
+  isBackup?: boolean;
+  onClick?: () => void;
+}
